Refactor code to make it more OOP (Minimal)
