+++ src/main/java/Command.java
@@ -1,0 +1,3 @@
+public abstract class Command {
+    public abstract CommandResult execute();
+}
