+++ src/main/java/CommandResult.java
@@ -1,0 +1,7 @@
+public class CommandResult {
+    public final String feedbackToUser;
+
+    public CommandResult(String feedbackToUser) {
+        this.feedbackToUser = feedbackToUser;
+    }
+}
